backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.2
python-jose[cryptography]==3.3.0
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.4 